# specs/1-phase-2-backend/contracts/openapi.yml
openapi: 3.0.0
info:
  title: Phase II Backend API
  version: 1.0.0
paths:
  /auth/signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User created successfully
  /auth/login:
    post:
      summary: Authenticate a user and get a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{task_id}:
    put:
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        is_completed:
          type: boolean
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum:
            - High
            - Medium
            - Low
        tags:
          type: array
          items:
            type: string
        recurrence:
          type: object
          properties:
            type:
              type: string
              enum:
                - daily
                - weekly
                - monthly
    TaskCreate:
      type: object
      properties:
        text:
          type: string
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum:
            - High
            - Medium
            - Low
        tags:
          type: array
          items:
            type: string
        recurrence:
          type: object
          properties:
            type:
              type: string
              enum:
                - daily
                - weekly
                - monthly
    TaskUpdate:
      type: object
      properties:
        text:
          type: string
        is_completed:
          type: boolean
        due_date:
          type: string
          format: date-time
        priority:
          type: string
          enum:
            - High
            - Medium
            - Low
        tags:
          type: array
          items:
            type: string
        recurrence:
          type: object
          properties:
            type:
              type: string
              enum:
                - daily
                - weekly
                - monthly
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT